Upload asm file

+++ p7_flask_app/segment.py
@@ -1,18 +1,21 @@
 from typing import Optional
 
+from config import config
+from p1_utils.domain import get_domain
 from p2_assembly.seg6_segment import Segment
 from p2_assembly.seg8_listing import LstCmd
 from p2_assembly.seg9_collection import SegLst, get_seg_collection
 from p4_execution.ex5_execute import TpfServer
 
 
-def reset_seg_assembly(blob_name: str) -> Optional[SegLst]:
+def reset_seg_assembly(blob_name: str, file_type) -> Optional[SegLst]:
     seg_collection = get_seg_collection()
-    seg_name = seg_collection.init_from_cloud(blob_name)
+    seg_name = seg_collection.init_from_cloud(blob_name, file_type)
     segment = seg_collection.get_seg(seg_name)
     if not segment:
         return None
-    LstCmd.objects.filter_by(seg_name=seg_name).delete()
+    if file_type == config.LST:
+        LstCmd.objects.filter_by(seg_name=seg_name).delete()
     SegLst.objects.filter_by(seg_name=seg_name).delete()
     seg: SegLst = get_seg_lst(segment)  # Assemble the segment and create LstCmd
     seg.create()
@@ -26,6 +29,7 @@ def get_seg_lst(segment: Segment) -> SegLst:
     seg_lst.file_type = segment.file_type
     seg_lst.source = segment.source
     seg_lst.blob_name = segment.blob_name
+    seg_lst.domain = get_domain()
     segment.assemble()
     seg_lst.error_line = str(segment.error_line)
     seg_lst.error_constant = segment.error_constant
